requirements.txt
certifi==2019.6.16
chardet==3.0.4
idna==2.8
numpy==1.17.0
Pillow==6.1.0
pydub==0.23.1
requests==2.22.0
urllib3==1.25.3
